Put validation in posting recipe & checked if recipe exits at the time of editing and deleting recipe

+++ api/recipe.js
@@ -2,6 +2,7 @@ const express = require('express');
 const router = express.Router();
 const Recipe = require('../models/recipe');
 const Ingredient = require('../models/ingredient');
+const validation = require('../validation');
 
 router.get('/', async(req, res)=>{
     try{
@@ -27,7 +28,6 @@ router.get('/', async(req, res)=>{
                 })
             };
         })
-        console.log(storeIngredientForRecipe)
         res.json(result)
     
         
@@ -45,40 +45,47 @@ router.post('/', async(req, res)=>{
         imageUrl: req.body.imageUrl,
         description: req.body.description,
         ingredients: req.body.ingredients
-    }    
-    try{
-        if (req.body.ingredients.length <= 0){         // recipe without ingrdient
-            recipeData = {
-                name: recipe.name,
-                imageUrl: recipe.imageUrl,
-                description: recipe.description
-            }
-            const createRecipe = await Recipe.create(recipeData)
-            res.status(200).json({mgs: "Recipe Created Succesfully"})
-        }else{                                              // recipe with ingredients
-            recipeData = {
-                name: recipe.name,
-                imageUrl: recipe.imageUrl,
-                description: recipe.description
-            }
-            const createRecipe = await Recipe.create(recipeData)
-            const allRecipeData = await Recipe.findAll();
-            let length = allRecipeData.length;
-            let index = allRecipeData[length-1].id;
-            for (let ingredient of recipe.ingredients){
-                ingredientData = {
-                    name: ingredient.name,
-                    ammount: ingredient.ammount,
-                    recipeId: index
-                }
-                const createIngredient = await Ingredient.create(ingredientData);
-            }
-            res.status(200).json({mgs: "Recipe & Ingredients Created Succesfully"});
-        }
-
-    }catch(err){
-        res.status(500).json({ err: err })
     }
+    try{
+        const validity = await validation.validateUser(recipe.name,recipe.imageUrl,recipe.description,recipe.ingredients);
+        try{
+            if (req.body.ingredients.length <= 0){         // recipe without ingrdient
+                recipeData = {
+                    name: recipe.name,
+                    imageUrl: recipe.imageUrl,
+                    description: recipe.description
+                }
+                const createRecipe = await Recipe.create(recipeData)
+                res.status(200).json({mgs: "Recipe Created Succesfully"})
+            }else{                                              // recipe with ingredients
+                recipeData = {
+                    name: recipe.name,
+                    imageUrl: recipe.imageUrl,
+                    description: recipe.description
+                }
+                const createRecipe = await Recipe.create(recipeData)
+                const allRecipeData = await Recipe.findAll();
+                let length = allRecipeData.length;
+                let index = allRecipeData[length-1].id;
+                for (let ingredient of recipe.ingredients){
+                    ingredientData = {
+                        name: ingredient.name,
+                        ammount: ingredient.ammount,
+                        recipeId: index
+                    }
+                    const createIngredient = await Ingredient.create(ingredientData);
+                }
+                res.status(200).json({mgs: "Recipe & Ingredients Created Succesfully"});
+            }
+    
+        }catch(err){
+            res.status(500).json({ err: err })
+        }
+    }catch(err){
+        res.json({msg: err})
+    }
+
+    
 })
 
 // For Edit Recipe
@@ -90,72 +97,81 @@ router.post('/:id', async(req,res) => {
         ingredients: req.body.ingredients
     } 
     try{
-        if(req.body.ingredients.length <= 0 ){           // only update the recipe with edit
-            recipeData = {
-                name: recipe.name,
-                imageUrl: recipe.imageUrl,
-                description: recipe.description
+        const findRecipeId = await Ingredient.findAll({
+            where: {
+                recipeId: req.params.id
             }
-            const updateRecipe = await Recipe.update(recipeData, {
-                where: {
-                  id: req.params.id
+        })
+        if(findRecipeId.length === 0){
+            res.status(500).json({ msg: "Please Check the Recipe!"})
+        }else{
+            if(req.body.ingredients.length <= 0 ){           // only update the recipe with edit
+                recipeData = {
+                    name: recipe.name,
+                    imageUrl: recipe.imageUrl,
+                    description: recipe.description
                 }
-              });
-            res.status(200).json({ msg: "Recipe Updated Successfully"});
-        }else{                                         // update with ingredients with edit
-            recipeData = {
-                name: recipe.name,
-                imageUrl: recipe.imageUrl,
-                description: recipe.description
-            }
-            const updateRecipe = await Recipe.update(recipeData, {
-                where: {
-                  id: req.params.id
-                }
-              });
-            const findRelatedIngredients = await Ingredient.findAll({
-                where:{
-                    recipeId: req.params.id
-                }
-            });
-            if(recipe.ingredients.length === findRelatedIngredients.length ){ //update if not ingredient add dynamically
-                let i = findRelatedIngredients[0].id;
-                for (let ingredient of recipe.ingredients) {
-                    if(i === findRelatedIngredients.length+1){
-                        break;
-                    }else{
-                        ingredientData  = {
-                            name: ingredient.name,
-                            ammount: ingredient.ammount
-                        }
-                        console.log(ingredientData)
-                        const updateIngredients = await Ingredient.update(ingredientData, {
-                            where: {
-                            recipeId: req.params.id,
-                            id: i     
-                            }
-                        });
-                        i++
+                const updateRecipe = await Recipe.update(recipeData, {
+                    where: {
+                      id: req.params.id
                     }
+                  });
+                res.status(200).json({ msg: "Recipe Updated Successfully"});
+            }else{                                         // update with ingredients with edit
+                recipeData = {
+                    name: recipe.name,
+                    imageUrl: recipe.imageUrl,
+                    description: recipe.description
                 }
-            }else{                                                       // update if ingredient add dynamically
-                const deleteData = await Ingredient.destroy({
+                const updateRecipe = await Recipe.update(recipeData, {
+                    where: {
+                      id: req.params.id
+                    }
+                  });
+                const findRelatedIngredients = await Ingredient.findAll({
                     where:{
                         recipeId: req.params.id
                     }
                 });
-                for (let ingredient of recipe.ingredients){
-                    ingredientData = {
-                        name: ingredient.name,
-                        ammount: ingredient.ammount,
-                        recipeId: req.params.id
+                if(recipe.ingredients.length === findRelatedIngredients.length ){ //update if not ingredient add dynamically
+                    let i = findRelatedIngredients[0].id;
+                    for (let ingredient of recipe.ingredients) {
+                        if(i === findRelatedIngredients.length+1){
+                            break;
+                        }else{
+                            ingredientData  = {
+                                name: ingredient.name,
+                                ammount: ingredient.ammount
+                            }
+                            const updateIngredients = await Ingredient.update(ingredientData, {
+                                where: {
+                                recipeId: req.params.id,
+                                id: i     
+                                }
+                            });
+                            i++
+                        }
                     }
-                    const createIngredient = await Ingredient.create(ingredientData);
+                }else{                                                       // update if ingredient add dynamically
+                    const deleteData = await Ingredient.destroy({
+                        where:{
+                            recipeId: req.params.id
+                        }
+                    });
+                    for (let ingredient of recipe.ingredients){
+                        ingredientData = {
+                            name: ingredient.name,
+                            ammount: ingredient.ammount,
+                            recipeId: req.params.id
+                        }
+                        const createIngredient = await Ingredient.create(ingredientData);
+                    }
                 }
+                
+                res.status(200).json({msg: "Recipe And Ingredients Updated Succesfully"})
             }
-            
-            res.status(200).json({msg: "Recipe And Ingredients Updated Succesfully"})
         }
+        
     }catch(err){
         res.status(500).json({err: err})
     }
@@ -163,19 +179,24 @@ router.post('/:id', async(req,res) => {
 
 router.post('/delete/:id', async(req,res) => {     
     try{
-        const deleteIngredient = await Ingredient.destroy({
-            where:{
-                recipeId: req.params.id
-            }
-        });
-        const deleteRecipe = await Recipe.destroy({
-            where:{
-                id: req.params.id
-            }
-        });
-        res.status(200).json({
-            msg: "Recipe Deleted Successfully"
-        })
+        const getId = await Recipe.findAll({where: {id: req.params.id}});
+        if(getId.length === 0){
+            res.status(500).json({msg: "Recipe Not Found"})
+        }else{
+            const deleteIngredient = await Ingredient.destroy({
+                where:{
+                    recipeId: req.params.id
+                }
+            });
+            const deleteRecipe = await Recipe.destroy({
+                where:{
+                    id: req.params.id
+                }
+            });
+            res.status(200).json({
+                msg: "Recipe Deleted Successfully"
+            })
+        }
     }catch(err){
         res.status(500).json({error: err});
     }
